rename Lannion Tregor 2024
