+++ api/src/pdc/services/policy/engine/policies/index.ts
@@ -7,7 +7,7 @@ import { GrandPoitiers } from './20230927_GrandPoitiers';
 import { Idfm } from './20210520_IDFM';
 import { LaRochelle20232024 } from './LaRochelle20232024';
 import { Lannion } from './20230501_Lannion';
-import { LannionTregor2024 } from './LannionTregor2024';
+import { LannionTregor2024 } from './20240401_LannionTregor';
 import { Laval } from './Laval';
 import { MetropoleSavoie } from './MetropoleSavoie';
 import { Montpellier } from './Montpellier';
